requirements.txt
Flask
gunicorn
azure-cognitiveservices-speech
openai
python-dotenv
fillpdf
azure-storage-blob